Create ROT-13.py
Method 1 - String ROT13
Encrypt user strings in ROT 13 encryption. ROT-13 is a simple substitution cypher. It stands for "ROTate by 13 places." Thus the cypher replaces any letter (a-z or A-Z) with the one that appears 13 sequential places behind it. 
For letters in the last half of the alphabet, their ROT-13 equivalents loop back around to the front of the alphabet. Any non-alphabetical characters simply pass through.